TRegex: Implemented profile-guided eager capture group matching

+++ graal-js/src/com.oracle.truffle.regex/src/com/oracle/truffle/regex/tregex/dfa/NFATransitionSet.java
@@ -165,14 +165,14 @@ public final class NFATransitionSet implements Iterable<NFAStateTransition> {
         if (target instanceof NFAMatcherState) {
             appendTransition(transition);
         } else if (target instanceof NFAAnchoredFinalState) {
-            anchoredFinalState = (short) target.getId();
+            anchoredFinalState = target.getId();
             if (target.hasPossibleResults()) {
                 updatePreCalcAnchoredResult(target.getPossibleResults().get(0));
             }
             appendTransition(transition);
         } else {
             assert target instanceof NFAFinalState;
-            finalState = (short) target.getId();
+            finalState = target.getId();
             if (target.hasPossibleResults()) {
                 updatePreCalcUnAnchoredResult(target.getPossibleResults().get(0));
             }
